+++ controllers/contacts/getAllContats.js
@@ -1,0 +1,12 @@
+const { listContacts } = require('../../models/contacts') ;
+
+const getAll =  async (req, res, next) => {
+    try {
+      const data = await listContacts();
+      res.json(data)
+    } catch (error) {
+      next(error);
+    }
+}
+
+module.exports = getAll;
